feat: Create Open Feign client module to modularise rest calls
Add necessary dependencies and configurations to enable feign clients
- Create clients module
- Create fraud client interface to store contracts for communicating with fraud service
- import clients module to customer service with 'compile' scope
- refactor customer service business logic to use FraudClient instead of restTemplate
- use requiredArgs constructor for dependency injection
- test manually on postman
