+++ clients/src/main/java/com/isikodon/clients/fraud/FraudCheckResponse.java
@@ -1,4 +1,4 @@
-package com.isikodon.fraud.model;
+package com.isikodon.clients.fraud;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
