+++ clients/src/main/java/com/isikodon/clients/fraud/FraudClient.java
@@ -1,0 +1,12 @@
+package com.isikodon.clients.fraud;
+
+import org.springframework.cloud.openfeign.FeignClient;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+
+@FeignClient("FRAUD")
+public interface FraudClient {
+
+    @GetMapping("api/v1/fraud-check/{customerId}")
+    FraudCheckResponse isFraudster(@PathVariable("customerId") Integer customerId);
+}
