+++ customer/pom.xml
@@ -9,7 +9,7 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>customerEntity</artifactId>
+    <artifactId>customer</artifactId>
 
     <properties>
         <maven.compiler.source>17</maven.compiler.source>
@@ -34,6 +34,11 @@
         <dependency>
             <groupId>org.springframework.cloud</groupId>
             <artifactId>spring-cloud-starter-netflix-eureka-client</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>com.isikodon</groupId>
+            <artifactId>clients</artifactId>
+            <version>1.0-SNAPSHOT</version>
         </dependency>
     </dependencies>
