+++ customer/src/main/java/com/isikodon/customer/CustomerMSApplication.java
@@ -3,9 +3,12 @@ package com.isikodon.customer;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
+import org.springframework.cloud.openfeign.EnableFeignClients;
 
-@SpringBootApplication
+
 @EnableEurekaClient
+@EnableFeignClients(basePackages = "com.isikodon.clients")
+@SpringBootApplication
 public class CustomerMSApplication {
     public static void main(String[] args) {
         SpringApplication.run(CustomerMSApplication.class, args);
