--- customer/src/main/java/com/isikodon/customer/model/FraudCheckResponse.java
@@ -1,12 +1,0 @@
-package com.isikodon.customer.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class FraudCheckResponse {
-    private boolean isFraudulent;
-}
